fixed formatting and a cast

+++ arch/X86/X86ATTInstPrinter.c
@@ -740,7 +740,7 @@ static void printOperand(MCInst *MI, unsigned OpNo, SStream *O)
 
 				if (opsize > 0)
 				{
-					MI->flat_insn->detail->x86.operands[MI->flat_insn->detail->x86.op_count].size = opsize;
+					MI->flat_insn->detail->x86.operands[MI->flat_insn->detail->x86.op_count].size = (uint8_t)opsize;
 					MI->flat_insn->detail->x86.encoding.imm_size = encsize;
 				}
 				else if (MI->op1_size > 0)
